2.2AB
removed volume menu for now, using Dmitry's button mods

Accent slashes reconfigured to use accelerometer.  Default is 2.2 can range from 2.0 to 4.0 based on taste

Spins trigger when a saber is spinning and has traversed 360 degrees.

Stabs trigger when a saber is held level and the forward acceleration is greater than 3.0